# DocMan_MCP/src/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "DocMan_MCP API",
    "description": "A comprehensive document management system with AI-powered summarization capabilities",
    "version": "1.0.0",
    "contact": {
      "name": "DocMan_MCP Team",
      "email": "support@docman-mcp.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.docman-mcp.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/documents/": {
      "get": {
        "summary": "List documents",
        "description": "Retrieve a list of all documents with pagination",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "description": "Number of records to skip",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of records to return",
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of documents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Document"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create document",
        "description": "Create a new document record",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Document": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "file_path": {
            "type": "string"
          },
          "owner_id": {
            "type": "integer"
          },
          "project_id": {
            "type": "integer",
            "nullable": true
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DocumentCreate": {
        "type": "object",
        "required": ["title", "file_path"],
        "properties": {
          "title": {
            "type": "string"
          },
          "file_path": {
            "type": "string"
          },
          "project_id": {
            "type": "integer",
            "nullable": true
          }
        }
      },
...
    }
  }
}